% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramounter-package.R
\docType{package}
\name{paramounter-package}
\alias{paramounter}
\alias{paramounter-package}
\title{paramounter: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

Authors:
\itemize{
  \item William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})
}

Other contributors:
\itemize{
  \item Jian Guo [copyright holder]
  \item Tao Huan [copyright holder]
}

}
\keyword{internal}
